Deleting BoorusScraper package adn consolidating to Boorus
