+++ src/DarkMage530/BulkImager/Boorus/DanbooruImage.java
@@ -1,4 +1,4 @@
-package DarkMage530.BulkImager.BooruScraper.Boorus;
+package DarkMage530.BulkImager.Boorus;
 
 import com.google.common.collect.Lists;
 import DarkMage530.BulkImager.Image;
@@ -10,15 +10,15 @@ import java.util.List;
 /**
  * Created by Shirobako on 10/1/2016.
  */
-public class ZerochanImage implements Image {
+public class DanbooruImage implements Image {
 
     private IqdbImage iqdbImage;
     private List<String> parsedTags;
     private File location;
 
-    public ZerochanImage(IqdbImage image) {
+    public DanbooruImage(IqdbImage image) {
         this.iqdbImage = image;
-        this.parsedTags = Lists.newArrayList(image.getRawTags().split(","));
+        this.parsedTags = Lists.newArrayList(image.getRawTags().split(" "));
     }
 
     public List<String> getTags() {
@@ -36,4 +36,5 @@ public class ZerochanImage implements Image {
     public String getRating() {
         return iqdbImage.getRating();
     }
+
 }
